backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
scikit-learn
prophet
python-multipart
plotly
kaleido
tensorflow
pillow
biopython